backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "henrylong", email = "henrylong719@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "email-validator>=2.3.0",
    "fastapi[standard]>=0.129.1",
    "pwdlib[argon2,bcrypt]>=0.3.0",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.11.0",
    "python-multipart>=0.0.22",
]

[project.scripts]
backend = "backend:main"

[build-system]
requires = ["uv_build>=0.9.17,<0.10.0"]
build-backend = "uv_build"
